codigo gerado pelo diagrama atualizado

+++ Produto.cpp
@@ -10,12 +10,12 @@ string Produto::getNome() {
 	throw "Not yet implemented";
 }
 
-int Produto::getPreco() {
+float Produto::getPreco() {
 	// TODO - implement Produto::getPreco
 	throw "Not yet implemented";
 }
 
-void Produto::changePreco(int Preco) {
+void Produto::changePreco(float Preco) {
 	// TODO - implement Produto::changePreco
 	throw "Not yet implemented";
 }
@@ -25,7 +25,7 @@ string Produto::getDescricao() {
 	throw "Not yet implemented";
 }
 
-Produto::Produto(double cod, string nome, int preco, string descricao) {
+Produto::Produto(double cod, string nome, float preco, string descricao) {
 	// TODO - implement Produto::Produto
 	throw "Not yet implemented";
 }
